+++ src/jsp/com/cubrid/plcsql/compiler/ParseTreeConverter.java
@@ -36,10 +36,10 @@ import com.cubrid.jsp.data.ColumnInfo;
 import com.cubrid.jsp.data.DBType;
 import com.cubrid.plcsql.compiler.antlrgen.PcsParserBaseVisitor;
 import com.cubrid.plcsql.compiler.ast.*;
-import com.cubrid.plcsql.compiler.serverapi.PlParamInfo;
-import com.cubrid.plcsql.compiler.serverapi.ServerAPI;
-import com.cubrid.plcsql.compiler.serverapi.ServerConstants;
-import com.cubrid.plcsql.compiler.serverapi.SqlSemantics;
+import com.cubrid.plcsql.compiler.PlParamInfo;
+import com.cubrid.plcsql.compiler.ServerAPI;
+import com.cubrid.plcsql.compiler.ServerConstants;
+import com.cubrid.plcsql.compiler.SqlSemantics;
 import java.math.BigDecimal;
 import java.math.BigInteger;
 import java.time.LocalDate;
